requirements.txt
requests==2.32.3
PyYAML==6.0.2
python-dotenv==0.19.2
feedparser==6.0.11
python-dateutil==2.9.0.post0
pytz==2024.1
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6

